[FIX] note only contains info of the current move
The communication given in a line of a coda file contains a lot of information (such as
the partner's name and home address, the bank's name and identification, the structured
communication...). This information is what we get in transaction_ref.
But in the move communication we only want to keep the actual payment transaction
communication, which is given in ref_move