[ADD] l10n_be_bpost_address_validation
This module allows you to check partner's addresses validity in one click.
If address is not valid a proposal change will be displayed and you will be free to keep it or not.
If you accept it, the partner's address will be automatically updated.

The validation of the address is done throught the bpost API available here: https://www.bpost.be/en/addressing-web-service-widgets-address-validation.
Bpost is Belgium's leading postal operator and plays a key role in maintaining the country's economic and social fabric.