[REF] sale_margin_tax: refactor for correct accounting
When accounting for margin tax sales, the purchase price
and margin (that form the base for the tax lines) need to be split.
This is because a different tax grid should be applied on both line.
However, Odoo does not provide a way to split the base as it does to
split the tax. Furthermore, adding a 'margin' line type does not work
because the functions that apply the repartition lines etc are basically
impossible to override (they do not provide any hook,
and the arbitrary compute order would make any tentative very brittle).
To avoid these difficulty, another line is created on a "margin" product
(a service), so that both the purchase and the margin can be accounted
as base lines.

This provide the advantage that no further override of basic accounting
functions are needed, and that the tax rate of the margin tax is always
either the initial rate (say 21) or 0.

The downside is that it makes so statistics harder.
Since the margin line and the product lines are linked by a many2one,
it is still possible to extract the relevant info, just a bit harder.

This refactoring also merges the margin behaviour from invoice and sale
into a mixin to avoid desync of the essentially duplicated code
(if only both line/tax fields had the same name...).
This solved a bug where an invoice would think to be margin-based but
without finding a witness margin tax.

Last, this now handles the case where the margin is negative,
in which case no tax is due.